+++ src/api/utils/isFull.js
@@ -1,4 +1,4 @@
-const isTeamPaid = require('./isTeamPaid');
+const hasTeamPaid = require('./hasTeamPaid');
 
 const isTeamFull = (team, max, paid = false) => {
   let count;
@@ -23,7 +23,7 @@ const isTournamentFull = async (tournament, req) => {
   }
   let teams = await Promise.all(tournament.teams.map(async (team) => {
     let isPaid = true;
-    isPaid = await isTeamPaid(req, team, null, tournament.playersPerTeam);
+    isPaid = await hasTeamPaid(req, team, null, tournament.playersPerTeam);
     return (isPaid ? 'paid' : 'empty');
   }));
   teams = teams.filter((team) => team !== 'empty');
